+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## 10th Feb 2025
+## Unreleased
 
 ### ft-sdk: 0.3.0
 
@@ -13,14 +13,16 @@
 - breaking: upgraded `ft_sys_common` to 0.2.0 (`ft-sys` re-exports
   `ft_sys_common::Email`)
 
+## 15th Feb 2025
+
 ### ft-sys-common: 0.2.0
 
 - breaking: `ft_sys_common::Email`
     - changed from `Vec<>` to `smallvec::SmallVec<>` for the fields
     - removed `subject`, `body_html` and `body_text` fields, content is managed
       via `ft_sys_common::EmailContent` now
-    - added `ft_sys_common::{EmailContent, RenderedEmail, SendEmailError}`
-    -
+    - added `ft_sys_common::{CancelEmailError, EmailContent, RenderedEmail, 
+      SendEmailError}`
 - added `ft_sys_common::Email::new()`
 - hidden docs for `host-only` feature
   `ft_sys_common::EmailHandle::{new, inner}()`